+++ lib/inspec-iggy/cloudformation/parser.rb
@@ -1,18 +1,17 @@
 # parses CloudFormation JSON files
 
-require 'json'
-
 require 'inspec/objects/control'
 require 'inspec/objects/ruby_helper'
 require 'inspec/objects/describe'
 
+require 'inspec-iggy/file_helper'
 require 'inspec-iggy/inspec_helper'
 
 module InspecPlugins::Iggy::CloudFormation
   class Parser
     # parse through the JSON and generate InSpec controls
     def self.parse_generate(file) # rubocop:disable all
-      template = parse_cfn(file)
+      template = InspecPlugins::Iggy::FileHelper.parse_json(file)
       absolutename = File.absolute_path(file)
 
       # InSpec controls generated
@@ -89,22 +88,6 @@ module InspecPlugins::Iggy::CloudFormation
       end
       Inspec::Log.debug "CloudFormation.parse_generate generated_controls = #{generated_controls}"
       generated_controls
-    end
-
-    # boilerplate JSON parsing
-    def self.parse_cfn(file)
-      Inspec::Log.debug "CloudFormation.parse_cfn file = #{file}"
-      begin
-        unless File.file?(file)
-          STDERR.puts "ERROR: #{file} is an invalid file, please check your path."
-          exit(-1)
-        end
-        JSON.parse(File.read(file))
-      rescue JSON::ParserError => e
-        STDERR.puts e.message
-        STDERR.puts "ERROR: Parsing error in #{file}."
-        exit(-1)
-      end
     end
   end
 end